Refactor and move helpers for Forwarded header E2E
* test/e2e/forwarded_header_policy_test.go (testRouteHeaders): Use new
buildCurlPod helper.
(TestForwardedHeaderPolicyAppend, TestForwardedHeaderPolicyReplace)
(TestForwardedHeaderPolicyNever, TestForwardedHeaderPolicyIfNone): Adjust
calls to buildEchoPod.
(buildEchoPod, buildEchoService, buildRoute): Move from here...
* test/e2e/util.go: ...to here.  New file.
(buildEchoPod): Add parameter for pod name.
(buildCurlPod): New function.
